Filled out the Makefile sorta did change-pos.
Note about `change-pos':

I'm not sure it works?  At least, in order to find out whether or not it works,
we'll have to test it in the full app.  The thing is that processes cannot
change the working directory of other processes, which includes their parent.
This means that running change-pos does nothing, as the dir change is undone the
second it returns.  Depending on how we run the programs from the shell, this
may not be an issue.